Add rejection chain functionality (#1)
* [WIP] Rejection chain

* Fix algorithm with rejection chain

* More fixes

* More fix

* Debug

* Add rejection chain functionality and debug helpers. Can now run runFromCurrent to get next stable matching (from MOSM to WOSM)

Implemented the basic functionality for rejection chains, so we can derive the next stable matching from the current one[1, 2]. 

The current implementation is not yet fully optimized. Repeated proposals are made. In addition, there are many restrictions in the member functions of the `Agent` and `Matching` class, namely many of them require the full preferences (and "(inverse) happinesses") are generated for all agents on both sides. Technically we can remove these constraints. Not doing so right now because it would be memory expensive to save all preferences when we only access a small portion of them, which would limit our ability to run simple one-pass experiments like MPDA on extra large markets.

We also made helpers for debugging such as `Matching::completePreferences`, `Matching::printAgentsPreferences`, `Matching::sanityCheckStableMatching`, `Agent::invHappinessFor `, `Agent::prefer`.

[1] https://arxiv.org/pdf/1910.04401.pdf (See Algorithm 2)
[2] https://web.stanford.edu/~iashlagi/papers/Unequalmatch_OnlineAppendix.pdf (See Algorithm 1)